+++ backend/src/main/java/com/overlang/domain/job/service/JobResultService.java
@@ -1,12 +1,16 @@
 package com.overlang.domain.job.service;
 
+import com.overlang.api.dto.job.CallbackOcrItemRequest;
 import com.overlang.api.dto.job.JobCallbackRequest;
+import com.overlang.api.dto.ocr.BlurRegionRequest;
 import com.overlang.api.dto.ocr.OcrItemResponse;
+import com.overlang.api.dto.ocr.OcrStyleRequest;
 import com.overlang.api.dto.segment.SegmentResponse;
 import com.overlang.api.dto.segment.SegmentWordResponse;
 import com.overlang.domain.job.entity.Job;
 import com.overlang.domain.job.repository.JobRepository;
 import com.overlang.domain.learning.repository.LearningContentRepository;
+import com.overlang.domain.ocr.entity.OcrItem;
 import com.overlang.domain.ocr.repository.OcrItemRepository;
 import com.overlang.domain.savedword.repository.SavedWordRepository;
 import com.overlang.domain.segment.entity.Segment;
@@ -158,5 +162,37 @@ public class JobResultService {
     }
 
     return segmentWords;
+  }
+
+  public void saveOcrItems(Job job, JobCallbackRequest request) {
+    if (request.ocrItems() == null) return;
+
+    List<OcrItem> ocrItems = request.ocrItems().stream().map(o -> toOcrItem(job, o)).toList();
+
+    ocrItemRepository.saveAll(ocrItems);
+  }
+
+  private OcrItem toOcrItem(Job job, CallbackOcrItemRequest o) {
+    OcrStyleRequest style = o.style();
+    BlurRegionRequest blurRegion = style != null ? style.blurRegion() : null;
+
+    return new OcrItem(
+        job,
+        o.startTime(),
+        o.endTime(),
+        o.originText(),
+        o.translatedText(),
+        o.boundingBox().x(),
+        o.boundingBox().y(),
+        o.boundingBox().w(),
+        o.boundingBox().h(),
+        o.confidence(),
+        style != null ? style.backgroundColor() : null,
+        style != null ? style.dominantBackgroundColor() : null,
+        style != null ? style.textColor() : null,
+        blurRegion != null ? blurRegion.x() : null,
+        blurRegion != null ? blurRegion.y() : null,
+        blurRegion != null ? blurRegion.w() : null,
+        blurRegion != null ? blurRegion.h() : null);
   }
 }
